+++ test/ReadonlyArray.ts
@@ -321,7 +321,7 @@ describe('ReadonlyArray', () => {
   })
 
   it('getEq', () => {
-    const O = _.getEq(S.Ord)
+    const O = _.getEq(S.Eq)
     U.deepStrictEqual(O.equals([])([]), true)
     U.deepStrictEqual(O.equals(['a'])(['a']), true)
     U.deepStrictEqual(O.equals(['a', 'b'])(['a', 'b']), true)
@@ -760,7 +760,7 @@ describe('ReadonlyArray', () => {
     }
 
     const eqA = pipe(
-      N.Ord,
+      N.Eq,
       Eq.contramap((f: A) => f.b)
     )
     const arrA: A = { a: 'a', b: 1 }
@@ -789,7 +789,7 @@ describe('ReadonlyArray', () => {
 
     assert.strictEqual(_.uniq(N.Eq)(_.empty), _.empty)
     const as: ReadonlyArray<number> = [1]
-    assert.strictEqual(_.uniq(N.Ord)(as), as)
+    assert.strictEqual(_.uniq(N.Eq)(as), as)
   })
 
   it('sortBy', () => {