+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 **Note**: A feature tagged as Experimental is in a
 high state of flux, you're at risk of it changing without notice.
 
+# 2.10.0-rc.5
+
 # 2.10.0-rc.4
 
 - **Deprecations**